Switch from adding inline style to adding attributes.
Earlier, I used the style attribute because of HTML's gradual move away
from attributes. I saw width and height style information and the attributes
as legacy support.

But it has become clear this is not the correct view. Resizing an image
client side should still be the job of CSS as that is a style issue. But
if we are wanting to provide width/height information to the browser that is
simple derived from the image (so the browser doesn't have to wait to download
and analyize the image) then that is content. Also see
[this StackOverflow](http://stackoverflow.com/a/2414940/120067) answer
for more detail on this interpretation of the spec.

So I believe the correct thing to do if you want to display a 100x100 image
as 50% it's size would be to have:

    <img src="/some/image.png" width="100" height="100"
      style="width: 50px; height: 50px">

This tells the browser it will be recieving an image that is 100x100 but that
it should resize that image to 50x50. This understanding should work with the
common responsive styling of letting the image resize based on it's container.
So your CSS would have:

    img {
      max-width: 100%;
      width: auto;
      height: auto;
    }

Your markup would be:

    <img src="/some/image.png" width="100" height="100">

This way the browser knows without downloading the image that the image
it is recieving will be 100x100. If the parent container is bigger than that
it will just reserve the 100x100 space so when the image is downloaded it
doesn't need to reflow the page. On the other hand, if the container is
smaller than 100x100 it can reserve just the space available (maintaining
aspect ratio based on the 100x100 information) and also insert the image when
downloaded without reflowing the page.